aggregate on AbstractTool
